+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update k8s.io dependencies to 0.18.9
+
 ## [2.0.1] - 2020-09-02
 
 ### Changed